+++ hpcladmin_flut/view-vendors.php
@@ -31,9 +31,9 @@ $Page = "View-Vendors";
 
 
 <?php
-if($_REQUEST["action"]=="delete")
+if (isset($_REQUEST["action"]) && $_REQUEST["action"] == "delete")
 {
-  $id = $_REQUEST["id"];
+  $id = $_REQUEST["id"] ?? '';
   $sql11 = "DELETE FROM tbl_users WHERE id = '$id'";
   $conn->query($sql11);
   ?>
